feat(Sidebar): insert sidebar component into MFE Sidebar
